+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to MiniJinja are documented here.
 
 ## 1.0.15
 
+- Resolved a compiler warning for Rust 1.77.  #440
+
 - Fixed an incorrect error case in `call_method`.  Now `UnknownMethod`
   is returned instead of `InvalidOperation` correctly. #439